+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Steroids Nest Changelog
 
+[Migration guide](docs/MigrationGuide.md)
+
+### Features
+- Переход с библиотеки `@ntegral/nestjs-sentry` на `@sentry/nestjs`
+
 ## [4.1.0](https://github.com/steroids/nest/compare/4.0.2...4.1.0) (2026-02-13)
 
 ### Features